BraveNewPipe: try to circumvent TransactionTooLargeException.
As workaround internally we use GZIP to reduce the amount of data that
is put as in intent.putExtra(). It has to be tested if that compressing
takes to much time and truncate here already.

As later on if we share on the ErrorActivity we have to truncate the traces
to around 100Kb as it otherwise crash there.