BraveNewPipeLegacy: use getApplicationContext to have a static variable inside BraveApp
In the App.java from main source set there is the static method App.getApp() from which
many parts of the application get ApplicationContext. But as the class App inherits from
BraveApp and sets the app variable in onCreate() only after the super.onCreate()
from BraveApp is called. Therefore App.getApp() has not yet an initialized variable
thus we need another way to get the ApplicationContext