BravePipe: Temporary additional check for new version until renaming is complete
To prepare the app for a future change of the update URL (as the project name)
has to change. We have to check both URL's in this release.

Also there is a check to the link of the projects homepage for this transition phase

This commit should be reversed in a later version.